# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Everise API",
    "version": "1.0.0",
    "description": "The API behind everise.dev. Generated from the routes (src/api)."
  },
  "paths": {
...
    "/api/articles": {
      "post": {
        "tags": [
          "articles"
        ],
        "summary": "Write a post",
        "security": [
          {
            "token": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewArticle"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The post.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleResponse"
                }
              }
            }
          },
          "default": {
            "description": "An error; `errors.body` says what went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "articles"
        ],
        "summary": "Posts, newest first, optionally filtered",
        "security": [
          {
            "token": []
          },
          {}
        ],
        "parameters": [
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "author",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "favorited",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "default": 20,
              "type": "integer",
              "minimum": -9007199254740991,
              "maximum": 9007199254740991
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "default": 0,
              "type": "integer",
              "minimum": -9007199254740991,
              "maximum": 9007199254740991
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticlesResponse"
                }
              }
            }
          },
          "default": {
            "description": "An error; `errors.body` says what went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/articles/{id}": {
      "get": {
        "tags": [
          "articles"
        ],
        "summary": "A post",
        "security": [
          {
            "token": []
          },
          {}
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The post.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleResponse"
                }
              }
            }
          },
          "default": {
            "description": "An error; `errors.body` says what went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "articles"
        ],
        "summary": "Change a post (its author only)",
        "security": [
          {
            "token": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The post.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleResponse"
                }
              }
            }
          },
          "default": {
            "description": "An error; `errors.body` says what went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "articles"
        ],
        "summary": "Delete a post (its author only)",
        "security": [
          {
            "token": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Done."
          },
          "default": {
            "description": "An error; `errors.body` says what went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/articles/{id}/favorite": {
      "post": {
        "tags": [
          "articles"
        ],
        "summary": "Favorite a post",
        "security": [
          {
            "token": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The post.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleResponse"
                }
              }
            }
          },
          "default": {
            "description": "An error; `errors.body` says what went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "articles"
        ],
        "summary": "Unfavorite a post",
        "security": [
          {
            "token": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The post.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleResponse"
                }
              }
            }
          },
          "default": {
            "description": "An error; `errors.body` says what went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/articles/feed": {
      "get": {
        "tags": [
          "articles"
        ],
        "summary": "Posts by the members the user follows",
        "security": [
          {
            "token": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "default": 20,
              "type": "integer",
              "minimum": -9007199254740991,
              "maximum": 9007199254740991
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "default": 0,
              "type": "integer",
              "minimum": -9007199254740991,
              "maximum": 9007199254740991
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticlesResponse"
                }
              }
            }
          },
          "default": {
            "description": "An error; `errors.body` says what went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tags": {
      "get": {
        "tags": [
          "articles"
        ],
        "summary": "The tags in use",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The tags.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagsResponse"
                }
              }
            }
          },
          "default": {
            "description": "An error; `errors.body` says what went wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "NewAttachment": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "image",
              "gif",
              "video"
            ]
          },
          "url": {
            "type": "string",
            "maxLength": 2000,
            "format": "uri"
          },
          "alt": {
            "type": "string",
            "maxLength": 200
          },
          "width": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10000
          },
          "height": {
            "type": "integer",
            "minimum": 1,
            "maximum": 10000
          }
        },
        "required": [
          "kind",
          "url"
        ]
      },
      "NewArticle": {
        "type": "object",
        "properties": {
          "article": {
            "type": "object",
            "properties": {
              "title": {
                "type": "string",
                "minLength": 1
              },
              "description": {
                "type": "string",
                "minLength": 1
              },
              "body": {
                "type": "string"
              },
              "tagList": {
                "type": "array",
                "items": {
                  "type": "string",
                  "minLength": 1
                }
              },
              "media": {
                "maxItems": 4,
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/NewAttachment"
                }
              }
            },
            "required": [
              "title",
              "description",
              "body"
            ]
          }
        },
        "required": [
          "article"
        ]
      },
      "ArticleUpdate": {
        "type": "object",
        "properties": {
          "article": {
            "type": "object",
            "properties": {
              "title": {
                "type": "string",
                "minLength": 1
              },
              "description": {
                "type": "string",
                "minLength": 1
              },
              "body": {
                "type": "string"
              },
              "tagList": {
                "type": "array",
                "items": {
                  "type": "string",
                  "minLength": 1
                }
              },
              "media": {
                "maxItems": 4,
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/NewAttachment"
                }
              }
            }
          }
        },
        "required": [
          "article"
        ]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "object",
            "properties": {
              "body": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "required": [
              "body"
            ],
            "additionalProperties": false
          },
          "unconfirmedEmail": {
            "type": "string"
          }
        },
        "required": [
          "errors"
        ],
        "additionalProperties": false
      },
...
      "Profile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "following": {
            "type": "boolean"
          },
          "bio": {
            "type": [
              "string",
              "null"
            ]
          },
          "image": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "following",
          "bio",
          "image"
        ],
        "additionalProperties": false
      },
...
      "Attachment": {
        "anyOf": [
          {
            "type": "object",
            "properties": {
              "kind": {
                "type": "string",
                "enum": [
                  "image",
                  "gif"
                ]
              },
              "url": {
                "type": "string"
              },
              "alt": {
                "type": "string"
              },
              "width": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "height": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              }
            },
            "required": [
              "kind",
              "url"
            ],
            "additionalProperties": false
          },
          {
            "type": "object",
            "properties": {
              "kind": {
                "type": "string",
                "const": "video"
              },
              "url": {
                "type": "string"
              },
              "videoId": {
                "type": "string"
              },
              "start": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "alt": {
                "type": "string"
              }
            },
            "required": [
              "kind",
              "url",
              "videoId"
            ],
            "additionalProperties": false
          }
        ]
      },
      "RouletteCard": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "mode": {
            "type": "string"
          },
          "dutyUnknown": {
            "type": "boolean"
          },
          "image": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              {
                "type": "null"
              }
            ]
          },
          "job": {
            "anyOf": [
              {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "icon": {
                    "type": "integer",
                    "minimum": -9007199254740991,
                    "maximum": 9007199254740991
                  }
                },
                "required": [
                  "name",
                  "icon"
                ],
                "additionalProperties": false
              },
              {
                "type": "null"
              }
            ]
          },
          "guest": {
            "type": "boolean"
          }
        },
        "required": [
          "type",
          "name",
          "detail",
          "mode",
          "dutyUnknown",
          "image",
          "job",
          "guest"
        ],
        "additionalProperties": false
      },
      "Article": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "tagList": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d:[0-5]\\d(?:\\.\\d+)?(?:Z))$"
          },
          "favorited": {
            "type": "boolean"
          },
          "favoritesCount": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "media": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Attachment"
            }
          },
          "roulette": {
            "$ref": "#/components/schemas/RouletteCard"
          },
          "author": {
            "$ref": "#/components/schemas/Profile"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "body",
          "tagList",
          "createdAt",
          "updatedAt",
          "favorited",
          "favoritesCount",
          "media",
          "author"
        ],
        "additionalProperties": false
      },
      "ArticleResponse": {
        "type": "object",
        "properties": {
          "article": {
            "$ref": "#/components/schemas/Article"
          }
        },
        "required": [
          "article"
        ],
        "additionalProperties": false
      },
      "ArticlesResponse": {
        "type": "object",
        "properties": {
          "articles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Article"
            }
          },
          "articlesCount": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "articles",
          "articlesCount"
        ],
        "additionalProperties": false
      },
...
      "TagsResponse": {
        "type": "object",
        "properties": {
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "tags"
        ],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "token": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "`Token <jwt>`, or the `token` cookie the site sets."
      }
    }
  }
}